Use MT19937 for pseudo-random number generations
PRNG is needed in grouping subgeneration in RAND GNC codes, as well
as in generating coding coefficients in non-binary precodes (i.e., during
building the bipartite graph). Previously we used the same PRNG as
ANSI-C, but unfortunately that PRNG has poor performance in terms of
period, which in particular affects GF(2) precoding for a large number
of source packets.

+++ examples/test.nhopRecoder.c
@@ -100,7 +100,7 @@ int main(int argc, char *argv[])
     for (i=0; i<numhop; i++) {
         if (argc == 16) {
             rate[i] = atoi(argv[14]);
-            pe[i] = atoi(argv[15]);
+            pe[i] = atof(argv[15]);
         } else if (argc == 14 + atoi(argv[13]) + 1 && atof(argv[15]) >= 1) {
             rate[i] = atoi(argv[14+i]);
             pe[i] = atof(argv[14+numhop]);
